Add a way to reference a lane independent of road driving side
A "Directed Lane" is an integer which counts lanes from the middle out, positive in the forward direction, negative in the backwards direction, regardless of which side of the road is driven on.

A LaneRef refers to a particular lane on a particular way and is used in Right of Way calculation in later commits.